renamed files that output jsx, fixed gatsby HEAD issue, moved contact form into it's own component
