+++ src/components/Gallery/constants/defaultImages.jsx
